Move logic to `.ts` file
